Refactor domain_coordinator API to use a context manager
The domain_coordinator refactored to expose a single context manager
which returns a unique ROS_DOMAIN_ID that is valid until the context
goes out of scope.

The function has been deliberately renamed to force fatal errors if
the call site was not updated to use a "with" context.

Fixes #9

Signed-off-by: Timo Röhling <timo.roehling@fkie.fraunhofer.de>