Ignore Ctrl-C in run_rmw_isolated on Windows (#24)
Like Linux, we want to ignore the signal in the wrapper process and let
it propagate to the child process. Note that we explicitly handle the
signal and do nothing rather than pass NULL to the function call. The
latter would propagate to the subprocess, which we don't want to happen.

Signed-off-by: Scott K Logan <logans@cottsay.net>