Stop loading RMW to load the test fixture
This change mirrors part of the RMW selection algorithm implemented in
rmw_implementation so that we can avoid actually loading the RMW just to
figure out what it's called.

The main reason we don't want to rely on the loaded RMW is that a
testing process may change the RMW_IMPLEMENTATION variable after
starting up with intent to invoke a subprocess that needs to be
isolated. We should assume that future communication will happen based
on the current environment variables, not the state when the process
started.

Signed-off-by: Scott K Logan <logans@cottsay.net>