Fix range for rmw_test_fixture_default port locking (#31)
This is an off-by-one error in the port locking range that can result in
a collision with domain_coordinator or colcon-ros-domain-id-coordinator.
The intent was to avoid using ROS_DOMAIN_ID=0. While we indeed avoid
that, the code wasn't locking the same port as the other mechanisms for
a given domain ID.

It should be 32768+ID, where because 0 is never used for testing, we
start allocation at 32769.

I also shrunk the top end to avoid some Linux-specific parts of that
range that we shouldn't be using. There are around 100 available ports
in this range.

Signed-off-by: Scott K Logan <logans@cottsay.net>