Completely unload RMW test fixture when stopped
The initiaial implementation of the RMW test fixture assumed that we
would never attempt to restart the fixture in the same process after
switching to a different RMW implementation. During rollout, I found
situations where we want to do exactly that.

This small change will cause the fixture discovery to re-run for
subsequent fixture starts after gracefully stopping the previous
fixture.

Signed-off-by: Scott K Logan <logans@cottsay.net>