Drop dependency group dependency on test fixtures (#60)
At the moment, all providers of test fixtures have implemented the
fixture in the same package which provides the RMW itself, meaning that
we don't need to use a dependency to ensure that the plugin is present
when the RMW is tested.

Having the dependency here means that all packages which implement a
plugin are upstream of this package, which inflates the dependency tree
in an undesirable way.

Signed-off-by: Scott K Logan <logans@cottsay.net>
(cherry picked from commit d346c033c7214739b03cf86bb8295fc893de43f3)